backend/requirements.txt
# Django and REST Framework
Django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# PDF Generation
reportlab>=4.0.0

# Other dependencies (add as needed)
# django-cors-headers>=4.0.0
# Pillow>=10.0.0

